requirements.txt
certifi==2019.9.11
chardet==3.0.4           
django==2.2.13
idna==2.8                 
pillow==7.0.0
pyparsing==2.4.6          
pytz==2019.3              
requests==2.22.0
six==1.14.0               
sorl-thumbnail==12.6.3
sqlparse==0.3.0           
urllib3==1.25.6           
zipp==2.2.0               
